+++ test/expression_test.rb
@@ -49,6 +49,26 @@ class ExpressionTest < Minitest::Test
     assert_equal(0, solver.value(x))
   end
 
+  def test_add_assumption
+    model = ORTools::CpModel.new
+    x = model.new_bool_var("x")
+    model.add_assumption(x)
+
+    solver = ORTools::CpSolver.new
+    assert_equal :optimal, solver.solve(model)
+    assert_equal true, solver.value(x)
+  end
+
+  def test_add_assumptions
+    model = ORTools::CpModel.new
+    x = model.new_bool_var("x")
+    model.add_assumptions([x])
+
+    solver = ORTools::CpSolver.new
+    assert_equal :optimal, solver.solve(model)
+    assert_equal true, solver.value(x)
+  end
+
   def test_to_s
     model = ORTools::CpModel.new
     x = model.new_int_var(0, 1, "x")